+++ otfunc.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <string.h>
 
-void VdWRaM(int NumbAT,float *VdWRad,float *LaVdWR)
+void VdWRaM(int NumbAT,double *VdWRad,double *LaVdWR)
 /* This function finds the largest van der Waals radius :
    LaVdWR  the largest van der Waals radius */
 {
@@ -17,14 +17,14 @@ void VdWRaM(int NumbAT,float *VdWRad,float *LaVdWR)
 
 
 
-void ReMMCo(int ReAtNu,float **ReCoor,float *ReMaxC,float *ReMinC)
+void ReMMCo(int ReAtNu,double **ReCoor,double *ReMaxC,double *ReMinC)
 /* This function finds the maximal and minimal coordinates of the receptor :
    ReMaxC  the maximal coordinates of the receptor (1->x,2->y,3->z)
    ReMinC  the minimal coordinates of the receptor (1->x,2->y,3->z) */
 {
   int i,j;
 
-/* Initialization */  
+/* Initialization */
   for (i=1;i<=3;i++) {
     ReMaxC[i]=ReCoor[1][i];
     ReMinC[i]=ReCoor[1][i];
@@ -37,16 +37,16 @@ void ReMMCo(int ReAtNu,float **ReCoor,float *ReMaxC,float *ReMinC)
         ReMaxC[j]=ReCoor[i][j];
       else if (ReCoor[i][j]<ReMinC[j])
         ReMinC[j]=ReCoor[i][j];
-    } 
+    }
   }
- 
+
 }
 
 
 
-void BSMMCo(float **ReCoor,int BSAtNu,int *BSAtLi,float *BSMaxC,float *BSMinC)
-/* This function finds the maximal and minimal coordinates of the binding 
-   site of the receptor : 
+void BSMMCo(double **ReCoor,int BSAtNu,int *BSAtLi,double *BSMaxC,double *BSMinC)
+/* This function finds the maximal and minimal coordinates of the binding
+   site of the receptor :
    BSMaxC  the maximal coordinates of the binding site (1->x,2->y,3->z)
    BSMinC  the minimal coordinates of the binding site (1->x,2->y,3->z) */
 {
@@ -58,7 +58,7 @@ void BSMMCo(float **ReCoor,int BSAtNu,int *BSAtLi,float *BSMaxC,float *BSMinC)
     BSMinC[i]=ReCoor[BSAtLi[1]][i];
   }
 
-/* Find the maximal and minimal coordinates of the binding site of the 
+/* Find the maximal and minimal coordinates of the binding site of the
    receptor */
   for (i=1;i<=BSAtNu;i++) {
     for (j=1;j<=3;j++) {
